// petforest-fe-cms/src/utils/enums/funeral/reservation-mngt.enums.ts
export enum DETAIL_PET_REGISTATION_OPTIONS {
  UNREGISTERED = 'UNREGISTERED',
  REGISTERED = 'REGISTERED',
}
export enum DETAIL_PET_TYPE_OPTIONS {
  DOG = 'dog',
  CAT = 'cat',
  BIRD = 'bird',
  HAMSTER = 'hamster',
  ETC = 'etc',
}
export enum DETAIL_PET_GENDER_OPTIONS {
  MALE = 'MALE',
  FEMALE = 'FEMALE',
  // OTHER = 'OTHER',
}
export enum DETAIL_GUARDIAN_BOOKING_OPTIONS {
  EASY = 'EASY',
  NAVER = 'NAVER',
  PHONE_CALL = 'PHONE_CALL',
  ANIMAL_HOSPITAL = 'ANIMAL_HOSPITAL',
  FRIEND = 'FRIEND',
  INTERNET = 'INTERNET',
  SNS = 'SNS',
}

export enum DETAIL_GUARDIAN_VISIT_OPTIONS {
  FRIEND = 'FRIEND',
  ANIMAL_HOSPITAL = 'ANIMAL_HOSPITAL',
  INTERNET = 'INTERNET',
  SNS = 'SNS',
  OTHER = 'OTHER',
}

export enum DETAIL_FUNERAL_RESERVATION_OPTIONS {
  RECEIPT = 'RECEIPT',
  CONFIRMED = 'CONFIRMED',
  CANCELLATION = 'CANCELLATION',
}

export enum DETAIL_FUNERAL_STATUS_OPTIONS {
  RESERVATION = 'RESERVATION',
  INPROGRESS = 'IN_PROGRESS',
  DEADLINE = 'DEADLINE',
  DEPARTURE = 'DEPARTURE_OF_TRANSPORT',
}

export enum DETAIL_VISTOR_NUMBER_OPTIONS {
  ONE_PERSON = 'ONE_PERSON',
  TWO_PERSON = 'TWO_PEOPLE',
  THREE_PERSON = 'THREE_PEOPLE',
  FOUR_PERSON = 'FOUR_AND_MORE',
}

export enum DETAIL_SERVICE_ATTENDANCE_OPTIONS {
  NON_APPEARANCE = 'NON_APPEARANCE',
  ATTENDANCE = 'ATTENDANCE',
}

export enum DETAIL_SERVICE_NUMBER_ATTENDANCE_OPTIONS {
  ONE_PERSON = 'ONE_PERSON',
  TWO_PEOPLE = 'TWO_PEOPLE',
  THREE_PEOPLE = 'THREE_PEOPLE',
}

export enum DETAIL_PAYMENT_METHODS_OPTIONS {
  CASH = 'CASH',
  ACCOUNT_AMOUNT = 'ACCOUNT_AMOUNT',
  CARD = 'CARD',
  // COMBINE = 'COMBINE',
  PRE_PAY = 'PRE_PAY',
}

export enum DETAIL_SERVICE_TRANSPORT_ACCOMPANIED_STATUS_OPTIONS {
  ACCOMPANY = 'ACCOMPANY',
  NOT_ACCOMPANY = 'NOT_ACCOMPANY',
}

export enum DETAIL_SERVICE_NEW_AUDITION_COLOR_OPTIONS {
  GOLD = 'GOLD',
  SILVER = 'SILVER',
}

export enum DETAIL_SERVICE_NEW_AUDITION_TYPE_OPTIONS {
  HOUSING_ENSHRINEMENT = 'monite',
  HOUSING_HOME = 'home',
  PRECIOUSNESS_MEMORY_WARMTH = 'emotional',
  ETERNITY_ELEGANCE = 'sophisticated',
  FRAME = 'frame',
  COLUMBARIUM = 'columbarium',
  MONITE = 'monite',
}

export enum DETAIL_SERVICE_LUCETTE_JEWELRY_OPTIONS {
  NOT_APPLIED = 'not_applied',
  APPLIED = 'applied',
}

export enum LUCETTE_PROCESS_STATUS {
  SCHEDULE_PENDING = 'SCHEDULE_PENDING',
  PROCESS_PENDING = 'PROCESS_PENDING',
  ON_THE_DAY = 'ON_THE_DAY',
  EXCLUSIVE = 'EXCLUSIVE',
}

export enum NOTI_HISTORY_CATEGORY {
  OSSUARY_CONTRACT_COMPLETED = 'OSSUARY_CONTRACT_COMPLETED',
  OBITUARY_PAYMENT_COMPLETED = 'OBITUARY_PAYMENT_COMPLETED',
  OBITUARY_PAYMENT_RECEIVING_MONEY = 'OBITUARY_PAYMENT_RECEIVING_MONEY',
  OSSUARY_MATURITY = 'OSSUARY_MATURITY',
  OSSUARY_EXTENSION_GUIDE = 'OSSUARY_EXTENSION_GUIDE',
  OSSUARY_EXTENSION = 'OSSUARY_EXTENSION',
  RESERVATION_ACCEPTANCE = 'RESERVATION_ACCEPTANCE',
  OBITUARY_LINK = 'OBITUARY_LINK',
  RESERVATION_CONFIRMED = 'RESERVATION_CONFIRMED',
  RESERVATION_ACCOMPANY_CONFIRMED = 'RESERVATION_ACCOMPANY_CONFIRMED',
  FUNERAL_HOME_PRODUCT_COMPLETED = 'FUNERAL_HOME_PRODUCT_COMPLETED',
  RESERVATION_MANAGER = 'RESERVATION_MANAGER',
  RESERVATION_MANAGER_CANCEL = 'RESERVATION_MANAGER_CANCEL',
  RESERVATION_CANCEL = 'RESERVATION_CANCEL',
  OBITUARY_GUEST_BOOK_REGISTRATION = 'OBITUARY_GUEST_BOOK_REGISTRATION',
  RESERVATION_NOT_ACCOMPANY_CONFIRMED = 'RESERVATION_NOT_ACCOMPANY_CONFIRMED',
  LUCETTE_CONFIRMED = 'LUCETTE_CONFIRMED',
  SATISFACTION_SURVEY = 'SATISFACTION_SURVEY',
  STOCK_ORDER_HOME = 'STOCK_ORDER_HOME',
  STOCK_ORDER_COLUMBARIUM = 'STOCK_ORDER_COLUMBARIUM',

  STOCK_ORDER_FRAME = 'STOCK_ORDER_FRAME',
  STOCK_ORDER_SOPHISTICATED = 'STOCK_ORDER_SOPHISTICATED',
  STOCK_ORDER_EMOTIONAL = 'STOCK_ORDER_EMOTIONAL',
}